Use (double,double) prototype for pow() as the previous use caused a problem when using cxx -strictansi on Tru64.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@46 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/mutlib/sp_alignment.cpp
@@ -595,7 +595,7 @@ int init_hash8n (
    if (word_length < 4) word_length = 4;
    if (word_length > 4) word_length = 8;
     }
-    size_hash = (int)pow(4.0, (float) word_length);
+    size_hash = (int)pow(4.0, (double) word_length);
     
     if ( HASH_JOB_BLKS & job ) {
    min_match = MAX(min_match,word_length);
